added echos and inputs

+++ mas.sh
@@ -1,13 +1,33 @@
 #!/bin/bash
 
-#Bitwarden
-mas install 1352778147
+echo "Do you want to install Bitwarden? (Y/n)"
+read -s -n 1 answer
+if [[ $answer == "y" || $answer == "" ]]
+then
+    mas install 1352778147
+fi
+echo ""
 
-#Telegram
-mas install 747648890
+echo "Do you want to install Telegram? (Y/n)"
+read -s -n 1 answer
+if [[ $answer == "y" || $answer == "" ]]
+then
+    mas install 747648890
+fi
+echo ""
 
-#Xcode
-mas install 497799835
+echo "Do you want to install Xcode? (Y/n)"
+read -s -n 1 answer
+if [[ $answer == "y" || $answer == "" ]]
+then
+    mas install 497799835
+fi
+echo ""
 
-#iStat Menus
-mas install 1319778037
+echo "Do you want to install iStat Menus? (Y/n)"
+read -s -n 1 answer
+if [[ $answer == "y" || $answer == "" ]]
+then
+    mas install 1319778037
+fi
+echo ""
